+++ lib/api/src/module.rs
@@ -3,6 +3,7 @@ use std::fmt;
 use std::fs;
 use std::io;
 use std::path::Path;
+use wasmer_types::ModuleHash;
 
 use crate::engine::AsEngineRef;
 use thiserror::Error;
@@ -367,7 +368,7 @@ impl Module {
     }
 
     /// Returns the xxhash module hash
-    pub fn hash(&self) -> [u8; 8] {
+    pub fn hash(&self) -> Option<ModuleHash> {
         self.0.hash()
     }
