+++ lib/types/Cargo.toml
@@ -23,6 +23,17 @@ enum-iterator = "0.7.0"
 target-lexicon = { version = "0.12.2", default-features = false }
 enumset.workspace = true
 bytecheck = "0.6.8"
+rand = "0.8"
+webc = { workspace = true }
+xxhash-rust = { version = "0.8.8", features = ["xxh64"] }
+sha2 = { version = "0.10" }
+hex = { version = "^0.4" }
+
+# `rand` uses `getrandom` transitively, and to be able to
+# compile the project for `js`, we need to enable this feature
+[dependencies.getrandom]
+version = "*"
+features = ["js"]
 
 [dev-dependencies]
 memoffset.workspace = true
