+++ lib/types/src/lib.rs
@@ -96,7 +96,7 @@ pub use crate::initializers::{
     OwnedDataInitializer, TableInitializer,
 };
 pub use crate::memory::{Memory32, Memory64, MemorySize};
-pub use crate::module::{ExportsIterator, ImportKey, ImportsIterator, ModuleInfo};
+pub use crate::module::{ExportsIterator, ImportKey, ImportsIterator, ModuleHash, ModuleInfo};
 pub use crate::units::{
     Bytes, PageCountOutOfRange, Pages, WASM_MAX_PAGES, WASM_MIN_PAGES, WASM_PAGE_SIZE,
 };
