+++ lib/types/src/module.rs
@@ -11,7 +11,10 @@ use crate::{
     LocalGlobalIndex, LocalMemoryIndex, LocalTableIndex, MemoryIndex, MemoryType, SignatureIndex,
     TableIndex, TableInitializer, TableType,
 };
+use std::fmt::{Display, Formatter};
+
 use indexmap::IndexMap;
+use rand::RngCore;
 use rkyv::{
     de::SharedDeserializeRegistry, ser::ScratchSpace, ser::Serializer,
     ser::SharedSerializeRegistry, Archive, Archived, CheckBytes, Deserialize as RkyvDeserialize,
@@ -19,6 +22,7 @@ use rkyv::{
 };
 #[cfg(feature = "enable-serde")]
 use serde::{Deserialize, Serialize};
+use sha2::Digest;
 use std::collections::BTreeMap;
 use std::collections::HashMap;
 use std::fmt;
@@ -97,8 +101,131 @@ mod serde_imports {
     }
 }
 
+/// The hash of a WebAssembly module.
+#[derive(
+    Debug,
+    Copy,
+    Clone,
+    PartialEq,
+    Eq,
+    Hash,
+    PartialOrd,
+    Ord,
+    RkyvSerialize,
+    RkyvDeserialize,
+    Archive,
+)]
+#[archive_attr(derive(CheckBytes, Debug))]
+#[cfg_attr(feature = "enable-serde", derive(Serialize, Deserialize))]
+pub enum ModuleHash {
+    /// xxhash
+    XXHash([u8; 8]),
+
+    /// sha256
+    Sha256([u8; 32]),
+}
+
+impl ModuleHash {
+    /// Create a new [`ModuleHash`] from the raw xxhash hash.
+    pub fn xxhash_from_bytes(key: [u8; 8]) -> Self {
+        Self::XXHash(key)
+    }
+
+    /// Create a new [`ModuleHash`] from the raw sha256 hash.
+    pub fn sha256_from_bytes(key: [u8; 32]) -> Self {
+        Self::Sha256(key)
+    }
+
+    /// Creates a random xxhash for the module
+    pub fn xxhash_random() -> Self {
+        let mut rand = rand::thread_rng();
+        let mut key = [0u8; 8];
+        rand.fill_bytes(&mut key);
+        Self::xxhash_from_bytes(key)
+    }
+
+    /// Creates a random sha256 hash for the module
+    pub fn sha256_random() -> Self {
+        let mut rand = rand::thread_rng();
+        let mut key = [0u8; 32];
+        rand.fill_bytes(&mut key);
+        Self::sha256_from_bytes(key)
+    }
+
+    /// Parse a XXHash hash from a hex-encoded string.
+    pub fn xxhash_parse_hex(hex_str: &str) -> Result<Self, hex::FromHexError> {
+        let mut hash = [0_u8; 8];
+        hex::decode_to_slice(hex_str, &mut hash)?;
+        Ok(Self::xxhash_from_bytes(hash))
+    }
+
+    /// Parse a Sha256 hash from a hex-encoded string.
+    pub fn sha256_parse_hex(hex_str: &str) -> Result<Self, hex::FromHexError> {
+        let mut hash = [0_u8; 32];
+        hex::decode_to_slice(hex_str, &mut hash)?;
+        Ok(Self::sha256_from_bytes(hash))
+    }
+
+    /// Generate a new [`ModuleCache`] based on the XXHash hash of some bytes.
+    pub fn xxhash(wasm: impl AsRef<[u8]>) -> Self {
+        let wasm = wasm.as_ref();
+
+        let hash = xxhash_rust::xxh64::xxh64(wasm, 0);
+
+        Self::XXHash(hash.to_ne_bytes())
+    }
+
+    /// Generate a new [`ModuleCache`] based on the Sha256 hash of some bytes.
+    pub fn sha256(wasm: impl AsRef<[u8]>) -> Self {
+        let wasm = wasm.as_ref();
+
+        let hash: [u8; 32] = sha2::Sha256::digest(wasm).into();
+
+        Self::Sha256(hash)
+    }
+
+    /// Get the raw hash.
+    pub fn as_bytes(&self) -> &[u8] {
+        match self {
+            Self::XXHash(bytes) => bytes.as_slice(),
+            Self::Sha256(bytes) => bytes.as_slice(),
+        }
+    }
+}
+
+impl From<webc::metadata::AtomSignature> for ModuleHash {
+    fn from(value: webc::metadata::AtomSignature) -> Self {
+        match value {
+            webc::metadata::AtomSignature::Sha256(bytes) => Self::Sha256(bytes),
+        }
+    }
+}
+
+impl Display for ModuleHash {
+    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
+        fn format<const N: usize>(f: &mut Formatter<'_>, bytes: &[u8; N]) -> fmt::Result {
+            for byte in bytes {
+                write!(f, "{byte:02X}")?;
+            }
+
+            Ok(())
+        }
+
+        match self {
+            Self::XXHash(bytes) => format(f, bytes)?,
+            Self::Sha256(bytes) => format(f, bytes)?,
+        }
+
+        Ok(())
+    }
+}
+
 /// A translated WebAssembly module, excluding the function bodies and
 /// memory initializers.
+///
+/// IMPORTANT: since this struct will be serialized as part of the compiled module artifact,
+/// if you change this struct, do not forget to update [`MetadataHeader::version`](crate::serialize::MetadataHeader)
+/// to make sure we don't break compatibility between versions.
 #[derive(Debug, Clone, Default)]
 #[cfg_attr(feature = "enable-serde", derive(Serialize, Deserialize))]
 pub struct ModuleInfo {
@@ -111,9 +238,8 @@ pub struct ModuleInfo {
     #[cfg_attr(feature = "enable-serde", serde(skip_serializing, skip_deserializing))]
     pub id: ModuleId,
 
-    // FIXME: We need ModuleHash here
     /// hash of the module
-    pub hash: [u8; 8],
+    pub hash: Option<ModuleHash>,
 
     /// The name of this wasm module, often found in the wasm file.
     pub name: Option<String>,
@@ -186,8 +312,7 @@ pub struct ModuleInfo {
 #[archive_attr(derive(CheckBytes, Debug))]
 pub struct ArchivableModuleInfo {
     name: Option<String>,
-    // FIXME: we need ModuleHash here
-    hash: [u8; 8],
+    hash: Option<ModuleHash>,
     imports: IndexMap<ImportKey, ImportIndex>,
     exports: IndexMap<String, ExportIndex>,
     start_function: Option<FunctionIndex>,
