+++ lib/wasix/src/runtime/module_cache/shared.rs
@@ -1,7 +1,8 @@
 use dashmap::DashMap;
 use wasmer::{Engine, Module};
 
-use crate::runtime::module_cache::{CacheError, ModuleCache, ModuleHash};
+use crate::runtime::module_cache::{CacheError, ModuleCache};
+use wasmer_types::ModuleHash;
 
 /// A [`ModuleCache`] based on a <code>[DashMap]<[ModuleHash], [Module]></code>.
 #[derive(Debug, Default, Clone)]
