+++ lib/wasix/src/runtime/module_cache/thread_local.rs
@@ -2,7 +2,8 @@ use std::{cell::RefCell, collections::HashMap};
 
 use wasmer::{Engine, Module};
 
-use crate::runtime::module_cache::{CacheError, ModuleCache, ModuleHash};
+use crate::runtime::module_cache::{CacheError, ModuleCache};
+use wasmer_types::ModuleHash;
 
 std::thread_local! {
     static CACHED_MODULES: RefCell<HashMap<(ModuleHash, String), Module>>
