+++ lib/wasix/src/runtime/module_cache/types.rs
@@ -1,13 +1,7 @@
-use std::hash::Hash;
-use std::{
-    fmt::{self, Debug, Display, Formatter},
-    ops::Deref,
-    path::PathBuf,
-};
+use std::{fmt::Debug, ops::Deref, path::PathBuf};
 
-use rand::RngCore;
-use sha2::Digest;
 use wasmer::{Engine, Module};
+use wasmer_types::ModuleHash;
 
 use crate::runtime::module_cache::FallbackCache;
 
@@ -125,108 +119,6 @@ pub enum CacheError {
 impl CacheError {
     pub fn other(error: impl std::error::Error + Send + Sync + 'static) -> Self {
         CacheError::Other(Box::new(error))
-    }
-}
-
-/// The hash of a WebAssembly module.
-#[derive(Debug, Copy, Clone, PartialEq, Eq, Hash, PartialOrd, Ord)]
-pub enum ModuleHash {
-    XXHash([u8; 8]),
-    Sha256([u8; 32]),
-}
-
-impl ModuleHash {
-    /// Create a new [`ModuleHash`] from the raw xxhash hash.
-    pub fn xxhash_from_bytes(key: [u8; 8]) -> Self {
-        Self::XXHash(key)
-    }
-
-    /// Create a new [`ModuleHash`] from the raw sha256 hash.
-    pub fn sha256_from_bytes(key: [u8; 32]) -> Self {
-        Self::Sha256(key)
-    }
-
-    /// Creates a random xxhash for the module
-    pub fn xxhash_random() -> Self {
-        let mut rand = rand::thread_rng();
-        let mut key = [0u8; 8];
-        rand.fill_bytes(&mut key);
-        Self::xxhash_from_bytes(key)
-    }
-
-    /// Creates a random sha256 hash for the module
-    pub fn sha256_random() -> Self {
-        let mut rand = rand::thread_rng();
-        let mut key = [0u8; 32];
-        rand.fill_bytes(&mut key);
-        Self::sha256_from_bytes(key)
-    }
-
-    /// Parse a XXHash hash from a hex-encoded string.
-    pub fn xxhash_parse_hex(hex_str: &str) -> Result<Self, hex::FromHexError> {
-        let mut hash = [0_u8; 8];
-        hex::decode_to_slice(hex_str, &mut hash)?;
-        Ok(Self::xxhash_from_bytes(hash))
-    }
-
-    /// Parse a Sha256 hash from a hex-encoded string.
-    pub fn sha256_parse_hex(hex_str: &str) -> Result<Self, hex::FromHexError> {
-        let mut hash = [0_u8; 32];
-        hex::decode_to_slice(hex_str, &mut hash)?;
-        Ok(Self::sha256_from_bytes(hash))
-    }
-
-    /// Generate a new [`ModuleCache`] based on the XXHash hash of some bytes.
-    pub fn xxhash(wasm: impl AsRef<[u8]>) -> Self {
-        let wasm = wasm.as_ref();
-
-        let hash = xxhash_rust::xxh64::xxh64(wasm, 0);
-
-        Self::XXHash(hash.to_ne_bytes())
-    }
-
-    /// Generate a new [`ModuleCache`] based on the Sha256 hash of some bytes.
-    pub fn sha256(wasm: impl AsRef<[u8]>) -> Self {
-        let wasm = wasm.as_ref();
-
-        let hash: [u8; 32] = sha2::Sha256::digest(wasm).into();
-
-        Self::Sha256(hash)
-    }
-
-    /// Get the raw hash.
-    pub fn as_bytes(&self) -> &[u8] {
-        match self {
-            ModuleHash::XXHash(bytes) => bytes.as_slice(),
-            ModuleHash::Sha256(bytes) => bytes.as_slice(),
-        }
-    }
-}
-
-impl From<webc::metadata::AtomSignature> for ModuleHash {
-    fn from(value: webc::metadata::AtomSignature) -> Self {
-        match value {
-            webc::metadata::AtomSignature::Sha256(bytes) => ModuleHash::Sha256(bytes),
-        }
-    }
-}
-
-impl Display for ModuleHash {
-    fn fmt(&self, f: &mut Formatter<'_>) -> fmt::Result {
-        fn format<const N: usize>(f: &mut Formatter<'_>, bytes: &[u8; N]) -> fmt::Result {
-            for byte in bytes {
-                write!(f, "{byte:02X}")?;
-            }
-
-            Ok(())
-        }
-
-        match self {
-            ModuleHash::XXHash(bytes) => format(f, bytes)?,
-            ModuleHash::Sha256(bytes) => format(f, bytes)?,
-        }
-
-        Ok(())
     }
 }
